Move 'renderer.render_data.RenderData' -> 'renderer.RenderData' and delete 'renderer.render_data'
