+++ src/main/scala/engine/render/renderer/Renderer.scala
@@ -1,7 +1,6 @@
 package engine.render.renderer
 
 import engine.render.window.Window
-import render_data.RenderData
 
 trait Renderer(val window: Window) {
