+++ src/main/scala/engine/render/renderer/render_element/RenderElement.scala
@@ -1,6 +1,6 @@
 package engine.render.renderer.render_element
 
-import engine.render.renderer.render_data.RenderData
+import engine.render.renderer.RenderData
 
 trait RenderElement {
   /*private[renderer]*/
